Moved to double precision + THD+N / THD improvement

+++ audio_record.cpp
@@ -153,7 +153,7 @@ bool audioRecorder::pause(bool p)
     return true;
 }
 
-void audioRecorder::get_data(std::vector<float>& data, size_t size)
+void audioRecorder::get_data(std::vector<double>& data, size_t size)
 {
     if (!m_ring_buffer){
         return;
@@ -164,9 +164,9 @@ void audioRecorder::get_data(std::vector<float>& data, size_t size)
 
     if (data.size() != size) data.resize(size, 0);
 
-    constexpr float inv16 = 1.0f / float(INT16_MAX);
+    constexpr double inv16 = 1.0 / double(INT16_MAX);
     for (int i = 0; i < size; ++i){
-        data[i] = float(read_buf[i]) * inv16;
+        data[i] = read_buf[i] * inv16;
     }
     m_ring_buffer->advance_read_ptr(fill_bytes);
 }
